+++ .slugger.json
@@ -1,0 +1,3 @@
+{
+  "slug-zip": true
+}
