+++ .travis.yml
@@ -2,6 +2,7 @@ language: node_js
 sudo: false
 node_js:
 - 4.3.2
+- 6.10.2
 env:
   global:
   - secure: kr5ulY4s2hnbqZLUJdgRFzYWVB3OKWPHvxbLRPXnOoSd2WaNUQPeCzDM407tqMG6O1hEXQmhpo/w/05rBtuwogGKFYihnrHDUZ5outYCzAn3hml3Sj6pAeN8SJx5fmjIO3zmcH0tNAMOxqyGT5lSVrD51q/AjVoO8d1YT3dyqZcx50lGr7IEipiKgVb0clq/X5E2RuzQVFHdTW8OZY53ZA0yt2Mm8VX5bKvFcW9N3KWNDs78iQLuD2e8j7aTv4Mokb34GvoIfYzdE2riIBFOLc0bFpypNQOeLLuLs9Oh+xrRzvYXkF3HjCaCHbpcLpXg1KQo/P0/eq7ggr1/xKVMup0uedWrwMYs75dsy/s8lunSQ+5QNy4y6wWUuZuLZzDKgma4LKSicevV4nqR148Vsoy62oLlnxeZjj4m+M/uYJhghNSmR570zHJjVwC7CDrmerCMnGueHq2CXyRWAceg9cmpyp19bYBcnSg96ce9nUqH3FkPHlvsXP4hC5PH+2SWU/7eskVYzpR83UPC1bydkJUxRdk0gG5IX7ylfgrSnM2mTpNgPCeToE0UDptBGqh2L+K0qJpxFSYKOVorpOQNRQwoetuE6PImEpa1M3rCRUljPYqbm9VRIxhBJFfg97Lbw3iLMr8ARVlZBmlyj8WVMgzoA84UvuFySxd/E+675fc=
