+++ package.json
@@ -1,7 +1,7 @@
 {
   "license": "BSD-2-Clause",
   "engines": {
-    "node": "4.3.2"
+    "node": "6.10.2"
   },
   "name": "patrol-rules-aws",
   "repository": {
@@ -13,35 +13,21 @@
   "bugs": {
     "url": "https://github.com/mapbox/patrol-rules-aws/issues"
   },
-  "jscsConfig": {
-    "preset": "airbnb",
-    "requireCamelCaseOrUpperCaseIdentifiers": null
-  },
   "version": "0.1.2",
   "dependencies": {
-    "app-root-path": "1.2.0",
-    "d3-queue": "2.0.3",
-    "lambda-cfn": "^1.0.0"
+    "d3-queue": "^3.0.0",
+    "@mapbox/lambda-cfn": "^2.0.0"
   },
   "scripts": {
     "test": "NODE_ENV=test tape test/*.test.js test/**/*.test.js",
-    "pretest": "eslint rules/* && jscs rules/*"
+    "pretest": "eslint **/*.js"
   },
   "devDependencies": {
-    "aws-sdk": "^2.2.35",
-    "nock": "7.5.0",
-    "tape": "^3.5.0",
-    "eslint": "^2.4.0",
-    "jscs": "^1.11.3",
-    "eslint-config-unstyled": "^1.1.0"
-  },
-  "main": "index.js",
-  "jshintConfig": {
-    "node": true,
-    "unused": true,
-    "globalstrict": false,
-    "undef": true,
-    "noarg": true
+    "aws-sdk": "^2.95.0",
+    "tape": "^4.0.0",
+    "eslint": "^4.0.0",
+    "eslint-config-unstyled": "^1.1.0",
+    "@mapbox/mock-aws-sdk-js": "^0.0.5"
   },
   "description": "AWS rules for Patrol"
 }
