+++ test/assumeRole.test.js
@@ -1,15 +1,13 @@
 var test = require('tape');
 
-var rule = require('../../rules/assumeRole');
+var rule = require('../assumeRole/function');
 var fn = rule.fn;
-var name = rule.config.name;
 
 test('assumeRole rule', function(t) {
-
   var event = {
     "detail": {
       "userIdentity": {
-        "userName": "bob",
+        "userName": "bob"
       },
       "requestParameters": {
         "roleArn": "arn:aws:iam::12345678901:role/Administrator-123456",
@@ -21,7 +19,7 @@ test('assumeRole rule', function(t) {
   process.env.disallowedRoles = 'Administrator, DBMaintenance';
 
   fn(event, {}, function(err, message) {
-    t.error(err, 'No error when calling ' + name);
+    t.error(err, 'No error when calling assumeRole');
 
     if (JSON.stringify(message.summary).match(/Disallowed role Administrator assumed by bob/)) {
       t.pass('Matches disallowed Administrator role');
@@ -34,7 +32,7 @@ test('assumeRole rule', function(t) {
   var event = {
     "detail": {
       "userIdentity": {
-        "userName": "bob",
+        "userName": "bob"
       },
       "requestParameters": {
         "roleArn": "arn:aws:iam::12345678901:role/basic-123456",
@@ -46,7 +44,7 @@ test('assumeRole rule', function(t) {
   process.env.disallowedRoles = 'Administrator, DBMaintenance';
 
   fn(event, {}, function(err, message) {
-    t.error(err, 'No error when calling ' + name);
+    t.error(err, 'No error when calling assumeRole');
     t.equal(message, 'Disallowed role was not assumed',
       'Does not match non disallowed role');
   });
@@ -61,7 +59,7 @@ test('assumeRole rule', function(t) {
   process.env.disallowedRoles = 'Administrator, DBMaintenance';
 
   fn(event, {}, function(err, message) {
-    t.error(err, 'No error when calling ' + name);
+    t.error(err, 'No error when calling assumeRole');
     t.equal(message, 'This is the error message',
       'errorMessage is returned in callback');
   });
