+++ test/disallowedResources.test.js
@@ -1,27 +1,24 @@
 var test = require('tape');
+var AWS = require('@mapbox/mock-aws-sdk-js');
 
-var rule = require('../../rules/disallowedResources.js');
+var rule = require('../disallowedResources/function.js');
 var fn = rule.fn;
-var name = rule.config.name;
 
-test('disallowedResources rule', function(t) {
+process.env.disallowedResourceArns = 'arn:aws:s3:::foo/bar/baz, arn:aws:s3:::foo/bar';
 
-  t.plan(18);
-
-  process.env.disallowedResourceArns = 'arn:aws:s3:::foo/bar/baz, arn:aws:s3:::foo/bar';
-
-  var event = {
-    "detail": {
-      "userIdentity": {
-        "userName": "bob",
-      },
-      "requestParameters": {
-        "roleArn": "arn:aws:iam::12345678901:role/Administrator-123456",
-        "roleSessionName": "bob"
-      }
+var eventFixture = {
+  "detail": {
+    "userIdentity": {
+      "userName": "bob"
+    },
+    "requestParameters": {
+      "roleArn": "arn:aws:iam::12345678901:role/Administrator-123456",
+      "roleSessionName": "bob"
     }
-  };
+  }
+};
 
+test('disallowedResources no matches', function(t) {
   var docNoMatch = {
     Statement: [
       {
@@ -32,14 +29,17 @@ test('disallowedResources rule', function(t) {
       },
     ]
   };
-
+  var event = eventFixture;
   event.detail.requestParameters.policyDocument = JSON.stringify(docNoMatch);
 
   fn(event, {}, function(err, message) {
-    t.error(err, 'No error when calling ' + name);
+    t.error(err, 'No error when calling function');
     t.deepEqual(message, [], 'No matched disallowed resources');
+    t.end();
   });
+});
 
+test('disallowedResources one statement one match', function(t) {
   var docMatch = {
     Statement: [
       {
@@ -53,19 +53,36 @@ test('disallowedResources rule', function(t) {
       },
     ]
   };
-
+  var event = eventFixture;
   event.detail.requestParameters.policyDocument = JSON.stringify(docMatch);
+
+  AWS.stub('IAM', 'simulateCustomPolicy', function(params, callback) {
+    var data = {
+      EvaluationResults: [
+        {
+          EvalResourceName: params.ResourceArns,
+          EvalDecision: 'allowed'
+        }
+
+      ]
+    };
+    callback(null, data);
+  });
 
   fn(event, {}, function(err, message) {
     t.equal(message.length, 1, 'There is only one result');
     t.equal(message[0].subject,
-      'Policy allows access to disallowed resources',
-      'Matches disallowed resources');
+            'Policy allows access to disallowed resources',
+            'Matches disallowed resources');
     t.equal(message[0].summary,
-      'Policy allows access to disallowed resources: arn:aws:s3:::foo/bar/baz, arn:aws:s3:::foo/bar',
-      'summary lists matched disallowed resources');
+            'Policy allows access to disallowed resources: arn:aws:s3:::foo/bar/baz, arn:aws:s3:::foo/bar',
+            'summary lists matched disallowed resources');
+    AWS.IAM.restore();
+    t.end();
   });
+});
 
+test('disallowedResources two statements one match', function(t) {
   var docMixed = {
     Statement: [
       {
@@ -86,7 +103,20 @@ test('disallowedResources rule', function(t) {
     ]
   };
 
+  var event = eventFixture;
   event.detail.requestParameters.policyDocument = JSON.stringify(docMixed);
+  AWS.stub('IAM', 'simulateCustomPolicy', function(params, callback) {
+    var data = {
+      EvaluationResults: [
+        {
+          EvalResourceName: params.ResourceArns,
+          EvalDecision: 'allowed'
+        }
+
+      ]
+    };
+    callback(null, data);
+  });
 
   fn(event, {}, function(err, message) {
     t.equal(message.length, 1, 'There is only one result');
@@ -95,9 +125,13 @@ test('disallowedResources rule', function(t) {
       'No matched disallowed resources');
     t.equal(message[0].summary,
       'Policy allows access to disallowed resources: arn:aws:s3:::foo/bar/baz, arn:aws:s3:::foo/bar',
-      'subjectFull lists matched disallowed resources');
+            'subject lists matched disallowed resources');
+    AWS.IAM.restore();
+    t.end();
   });
+});
 
+test('disallowedResources fuzzy match', function(t) {
   var docFuzzyMatch = {
     Statement: [
       {
@@ -117,16 +151,33 @@ test('disallowedResources rule', function(t) {
       }
     ]
   };
-
+  var event = eventFixture;
   event.detail.requestParameters.policyDocument = JSON.stringify(docFuzzyMatch);
+
+  AWS.stub('IAM', 'simulateCustomPolicy', function(params, callback) {
+    var data = {
+      EvaluationResults: [
+        {
+          EvalResourceName: params.ResourceArns,
+          EvalDecision: 'allowed'
+        }
+
+      ]
+    };
+    callback(null, data);
+  });
 
   fn(event, {}, function(err, message) {
     t.equal(message.length, 1, 'There is only one result');
     t.equal(message[0].subject,
-      'Policy allows access to disallowed resources',
-      'Matches fuzzy match S3 disallowed resources');
+            'Policy allows access to disallowed resources',
+            'Matches fuzzy match S3 disallowed resources');
+    AWS.IAM.restore();
+    t.end();
   });
+});
 
+test('disallowedResources two statements one fuzzy match', function(t) {
   var docKinesisMatch = {
     Statement: [
       {
@@ -147,8 +198,23 @@ test('disallowedResources rule', function(t) {
     ]
   };
 
+  var event = eventFixture;
   event.detail.requestParameters.policyDocument = JSON.stringify(docKinesisMatch);
   process.env.disallowedResourceArns = 'arn:aws:kinesis:us-east-1:123456789012:stream/foo-bar-KinesisStream-ABC*, arn:aws:s3:::foo/bar';
+
+  AWS.stub('IAM', 'simulateCustomPolicy', function(params, callback) {
+    var data = {
+      EvaluationResults: [
+        {
+          EvalResourceName: params.ResourceArns,
+          EvalDecision: 'allowed'
+        }
+
+      ]
+    };
+    callback(null, data);
+  });
+
 
   fn(event, {}, function(err, message) {
     t.equal(message.length, 1, 'There is only one result');
@@ -157,9 +223,13 @@ test('disallowedResources rule', function(t) {
       'Matches kinesis disallowed resources');
     t.equal(message[0].summary,
       'Policy allows access to disallowed resources: arn:aws:kinesis:us-east-1:123456789012:stream/foo-bar-KinesisStream-ABC*',
-      'subjectFull lists matched disallowed resources');
+            'subject lists matched disallowed resources');
+    AWS.IAM.restore();
+    t.end();
   });
+});
 
+test('disallowedResources three statements two matches', function(t) {
   var docTwoMatches = {
     Statement: [
       {
@@ -189,8 +259,22 @@ test('disallowedResources rule', function(t) {
     ]
   };
 
+  var event = eventFixture;
   event.detail.requestParameters.policyDocument = JSON.stringify(docTwoMatches);
   process.env.disallowedResourceArns = 'arn:aws:kinesis:us-east-1:123456789012:stream/foo-bar-KinesisStream-ABC*, arn:aws:s3:::foo/bar';
+
+    AWS.stub('IAM', 'simulateCustomPolicy', function(params, callback) {
+    var data = {
+      EvaluationResults: [
+        {
+          EvalResourceName: params.ResourceArns,
+          EvalDecision: 'allowed'
+        }
+
+      ]
+    };
+    callback(null, data);
+  });
 
   fn(event, {}, function(err, message) {
     t.equal(message.length, 1, 'There is only one result');
@@ -199,9 +283,13 @@ test('disallowedResources rule', function(t) {
       'Matches kinesis and s3 disallowed resources');
     t.equal(message[0].summary,
       'Policy allows access to disallowed resources: arn:aws:kinesis:us-east-1:123456789012:stream/foo-bar-KinesisStream-ABC*, arn:aws:s3:::foo/bar',
-      'subjectFull lists matched disallowed resources');
+            'subjectFull lists matched disallowed resources');
+    AWS.IAM.restore();
+    t.end();
   });
+});
 
+test('disallowedResources AccessDenied', function(t) {
   var event = {
     "detail": {
       errorCode: "AccessDenied",
@@ -210,9 +298,10 @@ test('disallowedResources rule', function(t) {
   };
 
   fn(event, {}, function(err, message) {
-    t.error(err, 'No error when calling ' + name);
+    t.error(err, 'No error when calling function');
     t.equal(message, 'This is the error message',
-      'errorMessage is returned in callback');
+            'errorMessage is returned in callback');
+    t.end();
   });
 
 });
